requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
PyYAML==6.0.1 